System and AI Changes; Bug Fix; Item Balance
AI IMPROVE
-AI will look at summon atk and dedicated damage to increase chances of using summon moves

NEW
-There is no longer a ULT move type, instead there is a tag saying if is ult, making ULTS able to miss or be dodged
-If an ULT misses or is dodged, the user gain +25% ult rate for 6 turns

FIX
-Tooltip would only show Sec if Shift was already pressed when hovered (AGAIN)
-Tooltip would keep same move until hover it again (AGAIN) (AGAIN)
-Armour Pen was not being counted or shown
-Jewel of the Druid healing was set to Sanity damage

ULTS
-Blood Pressure ULT -> MAGICAL
-Enchantment of the Deadly Shadows ULT -> MAGICAL
-Feiry Dance ULT -> MAGICAL
-Giga Impact ULT -> PHYSICAL
-Headshot ULT -> RANGED
-Let the Hunt Begin ULT -> STATMOD
-Mecha Suit ULT -> DEFENCIVE
-Windy Valley Tornado ULT -> MAGICAL

ITEMS
-Jewel of the Druid
Summon HP: (140% max mana) + (80%->76% magic power)
Summon ATK: (10%->6% max mana) + (40%->42% attack damage) NEW + (28% magic power)

-Shadow Shroom
Cd 14->13
Uses 2->3
stamina: 80->65 ++ (15% attack damage) over 3 turns
shield: 120 ++ (2% current health) shield over 3->4 turns
grant for 7->6 turns:
22%->24% movement speed

-Zap
+10->12 magic power
ACTIVE: deal 10->15 + (115% magic power)

-Prismatic Staff
PASSIVE
every 9->8 magic attacks
10->15 + (30% magic power) true damage

+++ Jogo/Game Project/Assets/Code/Utils.cs
@@ -10,7 +10,7 @@ public struct DMG
     public float healMana;
     public float healStamina;
     public int healSanity;
-    public int shield;
+    public float shield;
 
     public void Reset()
     {
